Various fixes

+++ src/AppBundle/Command/User/InitializeUserCommand.php
@@ -46,6 +46,9 @@ class InitializeUserCommand extends ContainerAwareCommand
         /** @var string $username */
         $username = $input->getOption("username");
 
+        /** @var bool $hasAllUserTenant */
+        $hasAllUserTenant = $this->getContainer()->getParameter("has_all_user_tenant");
+
         /** @var ?User $user */
         $user = $userManager->getOneBy(['username' => $username]);
         if ($user === null) {
@@ -69,12 +72,15 @@ class InitializeUserCommand extends ContainerAwareCommand
             $kibana->loadIndexPatternForUserTenant($user);
             $output->write(".");
 
-            $es->deleteIndex($user->getAllUserIndex());
-            $output->write(".");
-            $kibana->loadIndexPatternForAllUser($user);
-            $output->write(".");
-            $kibana->createAllUserDashboard($user);
-            $output->write(".");
+            if ($hasAllUserTenant === true) {
+                $es->deleteIndex($user->getAllUserIndex());
+                $output->write(".");
+                $kibana->loadIndexPatternForAllUser($user);
+                $output->write(".");
+                $kibana->createAllUserDashboard($user);
+                $output->write(".");
+            }
+
             $output->writeln("Done");
         }
